requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
mysqlclient==2.2.7
pillow==11.1.0
psycopg==3.2.6
psycopg2==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
stripe==12.0.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
